Refactor Code
Changed scope of Rank.GetConversionBracket() to private, as it isn't being used anywhere and is not necessary to be used anywhere else than in the constructor.
Moved some leftover Calculation Methods in the Rank Model into the Calculation Service for consistency. And maybe future Unit Tests. (lol)
Added some comments \o/